+++ autoelective/__init__.py
@@ -3,6 +3,6 @@
 # filename: __init__.py
 # modified: 2019-09-13
 
-__version__ = "2.1.1"
-__date__    = "2019.09.13"
+__version__ = "3.0.7"
+__date__    = "2020.02.18"
 __author__  = "Rabbit"