+++ autoelective/_internal.py
@@ -3,24 +3,15 @@
 # filename: _internal.py
 # modified: 2019-09-08
 
-__all__ = [
-
-    "mkdir",
-    "abspath"
-
-    "userInfo",
-]
-
 import os
-
-userInfo = {} # shared the user's custom options
-
 
 def mkdir(path):
     if not os.path.exists(path):
         os.mkdir(path)
 
-def abspath(*paths):
-    _BASE_DIR = os.path.dirname(__file__)
-    return os.path.normpath(os.path.abspath(os.path.join(_BASE_DIR, *paths)))
+def absp(*paths):
+    return os.path.normpath(os.path.abspath(os.path.join(os.path.dirname(__file__), *paths)))
 
+def read_list(file, encoding='utf-8-sig', **kwargs):
+    with open(file, 'r', encoding=encoding, **kwargs) as fp:
+        return [ line.rstrip('\n') for line in fp if not line.isspace() ]